Separate netstandard17 serialization tests into separate files
