+++ src/System.Collections/src/project.json
@@ -3,7 +3,7 @@
     "netstandard1.7": {
       "dependencies": {
         "Microsoft.NETCore.Platforms": "1.0.1",
-        "Microsoft.TargetingPack.Private.CoreCLR": "1.1.0-beta-24426-03"
+        "Microsoft.TargetingPack.Private.CoreCLR": "1.1.0-beta-24431-02"
       },
       "imports": [
         "dotnet5.8"
